Consolidate veripy into single-file implementation with lit tests
Merge py.py, ingestor.py, printer.py into veripy/main.py with a click
CLI (--ir, --dfy flags). Replace 40 pytest tests with 3 lit/FileCheck
tests and 3 pytest CLI tests. Add click, lit, filecheck dependencies.
