+++ tests/filecheck/.lit_test_times.txt
@@ -1,0 +1,3 @@
+1.308491e-01 abs.py
+1.296136e-01 constant.py
+1.297259e-01 if_else.py
