+++ tests/filecheck/lit.cfg
@@ -1,0 +1,5 @@
+import lit.formats
+
+config.name = "veripy"
+config.test_format = lit.formats.ShTest()
+config.suffixes = [".py"]
